--- .editorconfig
@@ -1,9 +1,0 @@
-root = true
-
-# Applies to all files
-[*]
-indent_style = space
-indent_size = 4
-end_of_line = lf
-charset = utf-8
-trim_trailing_whitespace = true
