+++ Definations/Bank.d.cs
@@ -2,17 +2,13 @@ using System.Text.Json.Serialization;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 
-namespace BudgetTracker.Defination
+namespace Defination;
+
+public class Bank : MongoObject
 {
-    public class Bank : MongoObject
-    {
-        [BsonElement("name")]
-        [JsonPropertyName("name")]
-        public string Name { get; set; } = "";
-    }
+    [BsonElement("name")]
+    [JsonPropertyName("name")]
+    public string Name { get; set; } = "";
 }
 
-namespace BudgetTracker.Injectors
-{
-    public interface IBankService : IMongoService<BudgetTracker.Defination.Bank> { }
-}
+public interface IBankService : IService<Bank> { }
