--- Definations/Security.d.cs
@@ -1,8 +1,0 @@
-namespace BudgetTracker.Defination
-{
-    public class HashDetails
-    {
-        public string Salt { get; set; } = "";
-        public string Password { get; set; } = "";
-    }
-}
