+++ Services/Transaction.service.cs
@@ -1,12 +1,11 @@
+using Defination;
 using MongoDB.Bson;
 using MongoDB.Driver;
 using System.Text.RegularExpressions;
-using BudgetTracker.Injectors;
-using BudgetTracker.Defination;
 
 using ListAPI = API.Transactions.List;
 
-namespace BudgetTracker.Services
+namespace Services
 {
     public class TransactionService : MongoServices<Transaction>, ITransactionService
     {
